Update to Scala 2.12 and latest libraries
Originally, I just was looking to update the dependency version of `scalajs-jquery` to be a 2.12 compatible version.  Looking at history it appears that the tutorial tends to be on the latest for libraries, and I figure it that eventually it will want switch to 2.12, so I updated everything.

I ran through the tutorial myself using these 2.12 and these versions.  I then switched to a clone on 2.11.8 and saw the new library versions also worked fine there.