Fix the computation of `bestScore` in an example. (#427)
The code in the example doesn't work since `Map.flatMap` returns a `Map` (on a tuple with 2 elements) so only the last score will be taken into consideration. The code below will return `("Jane", 200)` but it should be `("John", 1000)`

Making the flatMap return a list will return the list of tuple scores as intended.