Make it obvious how to contribute
I wanted to contribute and normally to do that I'd check out the content,
and get into some sort of feedback loop where I view the output, make
changes and iterate. I was unable to do this however as I am unfamiliar
with the Ruby/Jekyll ecosystem.

Based on the travis.yml file (it's great to have it!), I established
that as part of the set up we need to install RVM, set up bundler and
all the necessary dependencies.

I moved out the dependencies into the standard Gemfile approach of the
Ruby ecosystem.

This works on Ubuntu 18 and the Windows Subsystem for Linux.

Now I'm able to edit and see the impact of my changes immediately and
without problem :-)... and I hope the future contributor wil lbe able to
do the same.