Move Tutorials into Documentation
In my view, tutorials are definitely part of the documentation (eg think
Python: https://docs.python.org/3/tutorial/index.html ).

The structure will be clearer and you can navigate straight from a
tutorial back to the documentation.

Deleted unused 'boost' and 'firstapp' pages: it's better not to have
this sort of thing lying around (and possibly getting indexed by Google,
and when a user visits the page they find nothing).